Added navigation sidebar
New navigation sidebar gives quick access to all example-groups
mentioned in the spec. The sidebar scrolls to follow the user when it
leaves the viewport (less a small buffer). The sidebar has a show and
hide link to allow the user to toggle the sidebar away if more
horizontal space is required.